+++ test/test.sh
@@ -27,7 +27,6 @@ then
     #set -e
     # Python versions to test
     declare -a pythons=('python2.6' 'python2.7' 'python3.2' 'python3.3')
-    #declare -a pythons=('python3.2' 'python3.3')
 
     for i in "${pythons[@]}"
     do
@@ -42,9 +41,14 @@ then
         virtualenv --quiet -p $PYTHON ve$i || exit 1
         source ve$i/bin/activate 
         pushd ..
-        python setup.py install
+        if [ "$2" == "--quiet" ];
+        then
+            python setup.py --quiet install 2>&1 >/dev/null | grep error | grep -v Py | grep -v __pyx_
+        else
+            python setup.py install
+        fi
         popd 
-        pip install --quiet -r requirements.txt 
+        pip install --upgrade --quiet -r requirements.txt 
         echo  
         echo Starting tests in $i environment
         test nosetests-$i.xml